requirements.txt
Flask
mysql-connector-python
Flask-Bcrypt
Flask-Login
